pyproject.toml
[project]
name = "zoocache"
dynamic = ["version"]
description = "Cache that invalidates when your data changes, not when a timer expires. Rust-powered semantic invalidation for Python."
readme = "README.md"
license = { text = "MIT" }

authors = [
    { name = "Alberto Daniel Badia", email = "alberto_badia@enlacepatagonia.com" },
]

requires-python = ">=3.10"

dependencies = []

[project.optional-dependencies]
django = ["django>=5.2.11"]
telemetry-prometheus = ["prometheus-client>=0.20.0"]
telemetry-otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
telemetry = [
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]

[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "zoocache._zoocache"
python-source = "src"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
    "maturin>=1.7",
    "mkdocstrings>=1.0.1",
    "django>=5.2.11",
    "pytest-benchmark>=4.0.0",
    "djangorestframework>=3.16.1",
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-include-markdown-plugin>=7.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["zoocache"]
combine-as-imports = true
